move DAQ menu to separate TU
